YTA-3227 API: Link from annual statement to payment history and payment allocations
