--- resources/public/api/conf/1.0/schemas/PaymentHistoryWithAllocations.schema.json
@@ -1,139 +1,0 @@
-{
-  "type": "object",
-  "properties": {
-    "taxOfficeNumber": {
-      "$ref": "Definitions.schema.json#/definitions/taxOfficeNumber"
-    },
-    "taxOfficeReference": {
-      "$ref": "Definitions.schema.json#/definitions/taxOfficeReference"
-    },
-    "taxYear": {
-      "$ref": "Definitions.schema.json#/definitions/taxYear"
-    },
-    "payments": {
-      "type": "array",
-      "description": "Sequence of payments received in the requested tax year",
-      "items": {
-        "$ref": "#/definitions/paymentWithAllocations",
-        "minItems": 0
-      }
-    },
-    "_links": {
-      "type": "object",
-      "properties": {
-        "empRefs": {
-          "$ref": "Definitions.schema.json#/definitions/link"
-        },
-        "summary": {
-          "$ref": "Definitions.schema.json#/definitions/link"
-        },
-        "statements": {
-          "$ref": "Definitions.schema.json#/definitions/link"
-        },
-        "payments": {
-          "$ref": "Definitions.schema.json#/definitions/link"
-        },
-        "self": {
-          "$ref": "Definitions.schema.json#/definitions/link"
-        },
-        "next": {
-          "$ref": "Definitions.schema.json#/definitions/link"
-        },
-        "previous": {
-          "$ref": "Definitions.schema.json#/definitions/link"
-        }
-      },
-      "required": [
-        "empRefs",
-        "summary",
-        "statements",
-        "self"
-      ]
-    }
-  },
-  "required": [
-    "taxOfficeNumber",
-    "taxOfficeReference",
-    "taxYear",
-    "payments",
-    "_links"
-  ],
-  "definitions": {
-    "paymentWithAllocations": {
-      "type": "object",
-      "description": "A payment, with allocations, made by the client",
-      "title": "Description of a payment and allocations received",
-      "properties": {
-        "paymentDate": {
-          "description": "The date the payment was received",
-          "$ref": "Definitions.schema.json#/definitions/date"
-        },
-        "amount": {
-          "description": "The payment amount",
-          "type": "number"
-        },
-        "allocatedAmount": {
-          "description": "The amount from the original payment that has been allocated to charges",
-          "type": "number"
-        },
-        "method": {
-          "description": "Method of payment",
-          "type": "string",
-          "enum": [
-            "Debit Card",
-            "Cheque",
-            "BACS",
-            "CHAPS",
-            "Credit Card",
-            "Direct Debit",
-            "Online Payment",
-            "Bank Lodgement",
-            "Giro Receipts",
-            "Giro Credits",
-            "FPS Receipts",
-            "Internet Receipts",
-            "Girobank",
-            "Post Office",
-            "Nil Declarations",
-            "Paymaster",
-            "Voluntary Payments"
-          ]
-        },
-        "allocations": {
-          "type": "array",
-          "description": "Sequence of allocations made from the payment",
-          "items": {
-            "type": "object",
-            "properties": {
-              "taxYear": {
-                "$ref": "Definitions.schema.json#/definitions/taxYear"
-              },
-              "taxMonth": {
-                "$ref": "Definitions.schema.json#/definitions/taxMonth"
-              },
-              "amount": {
-                "description": "The allocation amount",
-                "type": "number"
-              },
-              "code": {
-                "$ref": "Definitions.schema.json#/definitions/chargeCode"
-              }
-            },
-            "minItems": 0,
-            "required": [
-              "taxYear",
-              "taxMonth",
-              "amount"
-            ]
-          }
-        }
-      },
-      "required": [
-        "paymentDate",
-        "amount",
-        "allocatedAmount",
-        "allocations"
-      ]
-    }
-  }
-}
