+++ test/common/Fixtures.scala
@@ -122,26 +122,6 @@ object Fixtures {
   def epayePaymentHistory(empRef: EmpRef, taxYear: TaxYear): String =
     s"""
        |{
-       |  "payments": [
-       |    {
-       |      "dateOfPayment": "${taxYear.yearFrom}-06-17",
-       |      "amount": "123.45"
-       |    },
-       |    {
-       |      "dateOfPayment": "${taxYear.yearFrom}-10-07",
-       |      "amount": "456.78"
-       |    },
-       |    {
-       |      "dateOfPayment": "${taxYear.yearFrom}-12-08",
-       |      "amount": "999.00"
-       |    }
-       |  ]
-       |}
-    """.stripMargin
-
-  def epayePaymentHistoryWithAllocations(empRef: EmpRef, taxYear: TaxYear): String =
-    s"""
-       |{
        |  "taxYear": {
        |    "yearFrom": ${taxYear.yearFrom}
        |  },
@@ -301,6 +281,9 @@ object Fixtures {
       |    },
       |    "previous": {
       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/statements/2016-17"
+      |    },
+      |    "paymentHistory": {
+      |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/2017-18"
       |    }
       |  }
       |}
@@ -361,54 +344,6 @@ object Fixtures {
   )
 
   def expectedPaymentHistoryJson(empRef: EmpRef, taxYear: TaxYear): JsValue = Json.parse(
-    s"""
-       |{
-       |  "taxOfficeNumber": "${empRef.taxOfficeNumber}",
-       |  "taxOfficeReference": "${empRef.taxOfficeReference}",
-       |  "taxYear": {
-       |    "year": "${taxYear.asString}",
-       |    "firstDay": "${taxYear.firstDay}",
-       |    "lastDay": "${taxYear.lastDay}"
-       |  },
-       |  "payments": [
-       |    {
-       |      "paymentDate": "${taxYear.yearFrom}-12-08",
-       |      "amount": 999.00
-       |    },
-       |    {
-       |      "paymentDate": "${taxYear.yearFrom}-10-07",
-       |      "amount": 456.78
-       |    },
-       |    {
-       |      "paymentDate": "${taxYear.yearFrom}-06-17",
-       |      "amount": 123.45
-       |    }
-       |  ],
-       |  "_links": {
-       |    "empRefs": {
-       |      "href": "/organisations/paye/"
-       |    },
-       |    "summary": {
-       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}"
-       |    },
-       |    "statements": {
-       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/statements"
-       |    },
-       |    "self": {
-       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.asString}"
-       |    },
-       |    "next": {
-       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.next.asString}"
-       |    },
-       |    "previous": {
-       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.previous.asString}"
-       |    }
-       |  }
-       |}
-     """.stripMargin
-  )
-
-  def expectedPaymentHistoryWithAllocationsJson(empRef: EmpRef, taxYear: TaxYear): JsValue = Json.parse(
     s"""
        |{
        |  "taxOfficeNumber": "${empRef.taxOfficeNumber}",
@@ -522,17 +457,17 @@ object Fixtures {
        |    "statements": {
        |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/statements"
        |    },
-       |    "payments": {
-       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.asString}"
+       |    "statement": {
+       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/statements/${taxYear.asString}"
        |    },
        |    "self": {
-       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.asString}/allocations"
+       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.asString}"
        |    },
        |    "next": {
-       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.next.asString}/allocations"
+       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.next.asString}"
        |    },
        |    "previous": {
-       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.previous.asString}/allocations"
+       |      "href": "/organisations/paye/${empRef.taxOfficeNumber}/${empRef.taxOfficeReference}/payment-history/${taxYear.previous.asString}"
        |    }
        |  }
        |}
